chore: move ButtonGroupTabKey enum
To AssetsFilterRow

Remove unused props in PortfolioCard
